+++ src/app/shared/shared.module.ts
@@ -1,16 +1,23 @@
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { FocusDirective } from './directives/focus.directive';
 import { HighlightDirective } from './directives/highlight.directive';
+import { OrderByPipe } from './order-by.pipe';
 
 @NgModule({
   declarations: [
     HighlightDirective,
-    FocusDirective
+    FocusDirective,
+    OrderByPipe
   ],
-  imports: [],
+  imports: [CommonModule, FormsModule],
   exports: [
     HighlightDirective,
-    FocusDirective
+    FocusDirective,
+    OrderByPipe,
+    CommonModule,
+    FormsModule
   ],
   providers: [],
 })