+++ changelog.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.4] - 2021-10-03
+### Added 
+- Added OrderBy pipe and used it in cartListComponent
+
+### Changed
+- Using currency and uppercase pipe in productComponent
+- Moved forms and common modules to shared
+- Now getting products and cart items using observable
 
 ## [0.0.3] - 2021-09-25
 ### Added